auth updated

+++ django_app/baseapp/views.py
@@ -1,24 +1,57 @@
 from django.shortcuts import render,redirect
-# from django.http import HttpResponse
+from django.http import HttpResponse
 from .models import Room,Topic
 from .forms import Roomform
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
+from django.contrib.auth import authenticate,login,logout
+from django.contrib.auth.forms import UserCreationForm
 
-def login(request):
+def loginpage(request):
+    page = "login"
+    if request.method=="POST":
 
-    # if request.method=="POST":
-    #     username=request.POST.get('username')
-    #     password=request.POST.get('password')
+        username=request.POST.get('username').lower()
+        password=request.POST.get('password')
 
-    #     try:
-    #         user = User.objects.get(username=username)
-    #     except:
-    #         messages.error(request, "Document deleted.")
+        try:
+            user = User.objects.get(username=username)
+            redirect('home')
+            
+        except:
+            messages.error(request, "Document deleted.")
 
-    context = {{'username':username}}
+        users = authenticate(request, username=username, password=password)
+        if users is not None:
+            login(request, users)
+            return redirect('home')
+        else:
+            messages.error(request, "wrong info ...")
+    context = {"page":page}
 
+    return render(request,'baseapp/login_register.html',context)
+
+def logoutpage(request):
+    logout(request)
+    return redirect('home')
+
+def registerpage(request):
+    page = "register"
+    form= UserCreationForm()
+
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            return redirect('home')
+        else:
+            messages.error(request,"an error acqure")
+
+    context = {"page":page, "form": form}
     return render(request,'baseapp/login_register.html',context)
 
 def home(request):
@@ -42,6 +75,7 @@ def room(request,pk):
     
     return render(request,'baseapp/room.html',{'room':room})
 
+@login_required(login_url='/loginpage')
 def createRoom(request):
     form = Roomform()
     if request.method == "POST":
@@ -52,9 +86,14 @@ def createRoom(request):
     context = {'form':form}
     return render(request,'baseapp/create_room.html',context)
 
+@login_required(login_url='/loginpage')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = Roomform(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed")
+
     if request.method == "POST":
     # why we add instance means for update in same room .
     # if we not added instance IT WILL CREATE NEW ROOM
@@ -68,6 +107,9 @@ def updateRoom(request,pk):
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
     
+    if request.user != room.host:
+        return HttpResponse("You are not allowed")
+
     if request.method == "POST":
         room.delete()
         return redirect('home')
